requirements.txt
uvicorn==0.29.0
fastapi==0.110.0
jinja2==3.1.3
python-multipart==0.0.9
websockets==12.0
Faker==23.2.1
loguru==0.7.2
starlette==0.36.3
python-socks==2.4.4